Resolving issues
Changes made:

Hyperlink colour changed to the standard hyperlink blue and underlined by default.

A text stroke (in same colour as link) was added to the links on hover for added indication that the link is selected with the cursor.

Fixed css file so footers aren't affected by changes to hyperlinks in page content

Headers were changed to emphasize the change from one section to the next.

Horizontal lines were added before each H2 in the rst files to help delineate sections within document pages.